dabac_rs: Add Rust-based LSM skeleton
These are the bare necessities to implement an LSM in Rust.

The LSM is implemented in security/dabac_rs/dabac.rs and uses the
bindings to C directly to set itself up. This is of course not
production ready, but enough to set up a minimal example of an LSM.

The lsm_info struct technically needs to be aligned, but Rust does not
allow this in combination with repr(transparent), so far this works.

The init and hook functions also work with name mangling, but since this
is the part that is very close to C, using them like this feels better
for some reason.

The unsafe Sync definitions should be ok because the C side uses these
pointers in the same way. The unsafe function call to security_add_hooks
is fine because nothing else is accessing the hooks in the meantime. The
unsafe to create the pointer to the file_permission call seems a bit
weird, since it only creates pointers and does not dereference them, so
that should be fine as well.

The C types need to be wrapped in structs that have implement Sync, so
that the static variables can be created. To avoid the most glaring
issues, the Opaque type is used.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>