dabac_rs: Add PAP
This adds a securityfs for dabac_rs to edit the policy and attributions.

The securityfs is implemented in C because the bindings for it in Rust
are missing, so this was a lot simpler. The C side only takes care of
registering the fs and provides thin wrapper functions which directly
call into Rust.

The Rust side then takes care of copying the data from userspace and
handles the updates (not implemented yet).

Linking the C side to the Rust side hides some unsafety, in C the
definitions for the glue functions exported from Rust contain a pointer
argument, while the same argument is represented as UserPtr (usize) in
Rust. Since there is no unsafe function to create a UserPtr from an
actual pointer, this is probably not the worst, but still a bit shady.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>