dabac_rs: Use Vecs as HashMap replacement
Since Rust for Linux currently offers no HashMap, use Vecs with usize as
"keys" instead. This has the advantage, that accesses are very fast, but
comes with the downside that the Vec has to be pre-allocated with the
maximum length allowed, raising memory requirements.

To somewhat limit the memory usage, the amount of possible attribute
identifiers is limited to 256 and the amount of possible users and
objects are limited to 4096.

This requires all entity identifiers to be usize or convertible to it.
Linux user ids are returned from the kernel as u32, which can be
converted to usize on all for testing relevant platforms. Instead of
using the filename, files are now identified using inodes, which are
already represented as usize.

The policy semantics change with this as well:
- attribute removals now only specify the id and remove them regardless
  of the value
- each attribute can only be set with one value, if a post-condition
  specifies to add an attribute already set, its value gets updated

Since attributions can either be set or unset, Rust's Option is used for
it. To not waste half of the memory for the discriminant, attribute
values are represented as NonZeroU32. If other values are needed, they
need to be encoded before the policy is loaded into the kernel.

The attribute stores are now locked at the beginning of policy
evaluation and only get unlocked after resolution is done, including
post-conditions. This is to prevent races where access decisions with
incompatible post-conditions both resolve their pre-condition as valid
and then both want to execute their port-condition.

The attribute stores are now public within the crate, since they get
locked by policy resolution in the PDP. For cleaner separation, this
could be done by a function in the PIP, but that does not change much.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>