dabac_rs: Differentiate between operations
Add operation identifiers to the policy so that different rules can be
defined for each operation. In this case here, the two different
operations are reading and writing files.

For each operation, multiple rules can be defined. This makes the policy
more interesting and more powerful, as otherwise only a single
post-condition would be possible for each operation. It does make the
data structures and policy resolution a bit more complex and inefficient
though, since now each rule of an operation needs to be evaluated, so
that the post-conditions for all valid rules can be executed.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>