dabac_rs: Add support for environmental attributes
Conceptually, environmental attributes allow policy rules to react to
changes in the environment, without requiring changes to the policy
itself or normal attributions.

As they are implemented, little differentiates them from user or object
attributes. Environmental attributes rely on a userspace service to
update them when they change. This is for simplicity in this
implementation. Some values could be set directly within the kernel, or
fetched on demand, but that is part of future work.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>