dabac_rs: Prevent errors from missing attributes
If an attribute that was not set is requested during a resolution
request, this would previously result in EINVAL getting generated. Since
this is not that helpful and it can be expected for a policy to deal
with new files, change this to default deny.

This change means in particular, that a pre-condition requiring "u4=o5",
where neither the user attribute with the id 4, nor the object attribute
with the id 5 are set, still evaluates to false, since a missing
attribute is treated as unfulfillable, similar to how NaN != NaN.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>