rust: hash: Enable const HashMap creation
With this HashMaps can be used in global locks. Since the hasher cannot
be initialized statically, the HashMap (the hasher/RandomState in
particular) needs to be properly initialized later before the HashMap is
used for the first time. The initialization actually just replaces the
whole HashMap with a new, properly initialized version and is threfore
safe to call.

This is a very quick and dirty edit to enable this, cutting many
corners, it just creates a seed set to 0 and a marker for an initialized
GlobalState, even though that has not been initialized. For the current
purposes this is ok, but not all that proper.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>