rust: hash: Remove common infallible allocations
For the most commonly used HashMap functions, remove the implementations
assuming infallible allocations or make them private. In this case this
targets the insert() and reserve() functions and instead adds the
insert_resize_if_needed and insert_if_capacity functions, which either
try to allocate additional memory, or only insert if there is still
capacity left.

More proper interfaces for the map would try the insertion based on the
underlying table and only return errors if an allocation was absolutely
necessary or only then try to allocate. With limited time, this suffices
though. Since this is somewhat quick and dirty again, it might break
some of the tests, but that's not too important here.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>