dabac_rs: Add resolution cache
This adds a cache similar to SELinux's AVC, it maps the access vector
(here: operation identifier, user id and inode) to the previously
calculated policy resolution. If an entry is found in the cache, policy
resolution immediately stops and the cached resolution is returned,
otherwise resolution continues as usual.

Since post-conditions change the resolution logic of the policy, any
post-condition getting executed invalidates the whole cache and resets
it. A more fine grained approach could improve this, but would require
storing many more details about the cached entries and also slow down
the cache reset phase.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>