dabac_rs: Add missing SAFETY comments
Not all of these are very sophisticated, especially the ones used in the
helpers are pretty weak or might even be flat out wrong. The ones in the
bindings should be somewhat decent, if a little brief. Of course most of
the functions may not only be called from C, but this is an easier wa to
phrase the requirements.

Technically the `read_str` and `update_policy_or_attrs`functions should
also be unsafe, since that's where the actual requirements regarding the
parameters are needed, but they are encapsulated within the module and
only used from the individual functions with the same requirements.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>