dabac_rs: bindings: Better reasoning for unsafe
Improve the reasoning and casts to access the security_hook_list. While
the way used before seemed to work, immediately casting to the first
value in the array seems a bit sketchy, so instead move the Opaque
inside the array to wrap the individual elements and then after getting
the pointer to the array from &raw mut first cast it to a slice to be
able to get the length and then cast it to the pointer to teh first
element, also removing the Opaque from the type.

This and getting the length from the pointer avoids creating temporary
references to the static mut and also removes the need for a constant
which contains the number of hooks. The gain of the later is marginal
though, the length still needs to manually be updated and now the check
if it is in bounds of i32 happens at runtime, so maybe this is the worse
option for that.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>