dabac_rs: bindings: Use interior mutability
Since `Opaque` contains an `UnsafeCell` and provides interior mutability
using it, stop using a static mut to store the `security_hook_list` and
instead use the interior mutability to provide write access to C during
initialization. This simplifies the security guarantees around unsafe
accesses quite a bit, since the code does not have to make sure anymore
that no references are created.

This also reverts some of the changes from the previous commit. To more
easily access the pointer to the array, move it back inside of the
`Opaque`. And to more easily get the amount of items in the list,
reintroduce the constant. Dynamically getting the length is possible,
but does not change the fact that it needs to be updated in the type of
the struct definition anyway.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>