dabac_rs: Use ArrayVec for post-conditions
Part of removing allocations from policy resolutions. Instead of storing
post-conditions in an array, use an arrayvec for easier handling and
increased performance (no options, fewer checks).

This limits the amount of post-conditions that can be executed for one
operation, the amount is set as a constant in the crate root. It is
currently set to 32, which should be plenty.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>