dabac_rs: Get an AC decision before policy updates
Ask the PDP from the PAP when the policy gets read or updated if that
access should be allowed. This does not really fit the current
semantics, where objects are files only, since the object here is
actually either the policy in total or part of it (policy/attributions).
It does however somewhat fit as long as the rules do not require object
attributes in their pre-conditions.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>