dabac_rs: Remove initial cache implementation
Its semantics are wrong, post-conditions still need to be executed, even
if the policy result can be determined from the cache. To track, which
post-conditions need to be executed, they either need to be added to the
cache (difficult without them having identifiers), or the caching needs
to be done one the level of pre-conditions.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>