dabac_rs: Limit allocations by pre-allocating
During policy resolution, if rules require access to attribute
identifiers outside of the range of the currently allocated size of the
Vec, it needs to be lengthened first. To prevent this, always
pre-allocate the Vecs to the maximum identifier the policy requires.

This still does not prevent all allocations. If the post-conditions set
attributes for previously non-existing subjects or objects, those need
to be allocated regardless.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>