dabac_rs: Add top-level caching
First caching variant, cache requests at the top level.

This still needs to copy the environmental attributes each time values
are stored in the cache. This could be prevented by storing them in
Arcs. As long as is only a limited amount of them, they could also be
stored in const generic arrays or another without an allocation, Arcs
are more future-proof and realistic though.

Because the cache is not hashmap-based, it does not detect duplicates,
so those need to be detected by the code managing the cache. For this,
the cached variable stores if the entry was found in the cache. The
entry variable cannot be used directly, as that would result in multiple
mutable borrows to be active at the same time.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>