dabac_rs: Use hashing for the cache keys
This avoids copies of the environmental attributes vector while keeping
the security guarantees. The hashing uses SHA-256 through the kernel
crypto API. SHA-512/256 would likely be faster, but is not available.

For global caching this is unlikely to be beneficial, so this serves
more as a proof of concept for formula level caching.

The Attributes::as_bytes() function was checked using miri and reported
no issue.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>