dabac_rs: Split into 5 LSMs for caching variants
To be able to evaluate the possible caching variants from a single
kernel, split the dabac_rs LSM into 5 different variants. They all
protect the same files in /home/dabac_rs, so for evaluation only one
should be enabled at a time using the SECURITY kernel command line
parameter.

To not duplicate the whole code, but only the parts necessary (mainly
because of constants and unmangled functions), these parts are moved to
folders for the variants, while everything else is reused. For this, the
different variants have their own entrypoints (for example
dabac_tlc.rs), where the correct variant files are used. The C code for
the PAP bindings is linked depending on the variant.

To make sure multiple LSMs are not competing for the same directory and
files in the securityfs, the LSM signals to the fs implementation that
it was initialized. The securityfs checks that and only loads if the LSM
was loaded before. This check is implemented using a global variable,
same as AppArmor does it. The mutable static storing the LSM loaded
state is only used in C so that it does not need to uphold Rust's
guarantees for that, Rust only sees an FFI call.

The currently planned variants are:
- nc => no caching
- tlc => top level caching
- tlch => top level caching +hashing
- flc => formula level caching
- flch => formula level caching +hashing

At the moment only the TLCH variant is available, because that was
implemented last.

The variants will double again when the ones for locking (Mutex vs
Spinlock) are added.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>