dabac_rs: Add FLC variant
Add the necessary files to use the formula level caching (without
hashing) variant. The PDP for this variant does not use hashing and
instead stores the operationid, ruleid, user attributes, object
attributes and environmental attributes in the cache directly. Copying
those all the time is likely very inefficient, so this is mostly for
comparison.

One drawback of this architecture, is that it does not detect when
pre-conditions are the same, so it does not reuse cache entries even
when it could. For this there needs to either be a unique pre-condition
identifer or the formula itself needs to be hashed.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>