dabac_rs: Add FLCH variant
Add the necessary files to use the formula level caching with hashing
variant. The PDP for this variant hashes the operationid, ruleid, user
attributes, object attributes and environmental attributes with SHA-256
using the kernel crypto API to avoid additional copies and allocations
for the attributions.

The same drawback as for the FLC variant applies, it also does not
detect when pre-conditions are the same.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>