dabac_rs: Add spinlock variants
The amount of duplicated code is becoming unwieldy, now there are ten
different variants (2 locking variants * 5 caching variants).

Luckily the only thing left to change/add is the timestamp collection.

With the preparations, adding the spinlock variants is rather
straightforward, duplicate all the code, fix the names so that
everything links, change the backend in the sl variants.

Variant identifiers:
- mx: using mutexes
- sl: using spinlocks

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>