abac: Add prefix to duplicate symbols
In order to evaluate the LSMs, make sure all of them can be compiled
into the kernel together (even if they cannot be loaded at the same
time because they occupy the same files in the securityfs). The
evaluated LSM can then be chosen using the SECURITY parameter on the
kernel command line.

This is only the mechanical changes, renaming all duplicate symbols by
adding the correct prefix (abac_{rules,trees}{,_enc}).

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>