abac: Adapt to current LSM registration
Since the LSM was created, the LSM registration mechanism was changed to
be more static, so add all the necessary definitions for the ABAC LSM(s)
and change the registration to use the new process with a separate
lsm_id struct.

The MAX_LSM_COUNT solution is not super nice because the ABAC LSM
actually defines four different LSMs, but all four have the same "abac"
name and cannot be used at the same time because they all share many
symbol names, so this works decently well enough.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>