dabac_rs: Use individual paths in securityfs
To fix the last blocker for the LSMs actually working together, choose a
unique path where each of the variants places its files in the
securityfs, for example dabac_rs_mx_nc. If only one variant is loaded,
the generic dabac_rs directory is used in addition to simplify the
evaluation, since the code then does not have to distinguish between the
variants for that.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>