security: Increase maximum possible LSM count
For their static calls, the MAX_LSM_COUNT is determined at compile time.
Previously the macros calculating the amount and unrolling the function
definitions were limited to 15/20 respectively.

With the 10 possible dabac_rs variations, this is not enough anymore. To
make sure all LSMs could theoretically still be enabled, expand both the
UNROLL and the COUNT_ARGS macros to 30.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>