dabac_rs: Allow variants to be used concurrently
Define constants for each dabac_rs variant in the LSM counting mechanism
to make sure they can all be used together, now that the securityfs
implementations do not compete for the same paths anymore.

To make sure this works the ARGS_COUNT and UNROLL macros had to be
extended to allow larger numbers.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>