dabac_rs: Measure performance using cycle counts
Add performance measurements using processor cycle counts (rdtscp).
Depending on the variant, not all values can be measured, these will
default to 0, which can be detected from user space. Depending on the
kernel config, either no measurement is taken (for tests from user
space), only the start and the end are measured (for total performance)
or intermediate values are measured as well (for detailed analysis).
These are individual measurements so that they do not influence each
other.

The start of the performance measurements uses an lfence instruction in
addition to to reading the timestamp count to ensure no benchmarked
instruction starts execution beforehand. There is conflicting advice
whether the fence should be added before or after reading the count. All
other times are only measured using the rdtcsp instruction, since it
waits for all previous instructions to finish.

The performance measurement functions are implemented so that they have
the least possible visual impact on the surrounding code, which means
config flags are checked inside the functions.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>