abac: Measure performance using cycle counts
Add performance measurements using processor cycle counts (rdtscp) the
same way dabac_rs measures its performance. Also depending on the
variant, some values are not measured, which also default to 0.

The ABAC variants use the same config values to determine if
intermediate values should be recorded.

The C code to collect the timestamps should have the same functionality
as the Rust code, using an lfence for the start of the measurement and a
single rdtscp for everything else.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>