abac: Load a small example policy on startup
To simplify small tests using the LSM, load a small initial policy that
covers basic access, like dabac_rs does.

This is also to mitigate that something within qemu splits the write
syscall to the securityfs into multiple ones, one for each line, which
completely breaks the parsing and memory management code in very fun and
totally not kernel panicking ways. On the eval system it works so far
for smaller policies, let's see if eval ones break this.

Interestingly, the Rust implementation receives all data in a single
syscall, as it also is not written to handle partial writes.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>