dabac_rs: Store perf results in the kernel
Instead of printing them to dmesg, store the results of the in-kernel
performance measurements in a buffer, indexed by the userspace process
making the request. That way requests can be matched afterwards.

Getting the data is similar to the ABAC LSM, a securityfs entry controls
whether perf values are stored and can be used to read the data later.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>