abac: Provide perf results through fs like dabac
The ABAC LSM already did provide a way to get performance data through
its securitfyfs, but that was for the time based evaluation of only the
diff value. To instead provide json formatted data of all timestamps,
with the same format dabac_rs uses (for simpler handling in the perf
orchestrator), provide a json encoding function and change the perf
interface to mirror dabac_rs.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>