dabac_rs: Do not check policy in PAP during eval
If the kernel is configured for LSM performance measurements, do not
check for permissions when the policy gets updated. This is to make sure
that all policies for evaluation can be loaded no matter what rules they
define. The evaluation policies could be modified to always include
permissions for root to change the policy, but this is simpler and makes
it less likely the kernel gets locked accidentally by including a policy
that cannot be removed anymore.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>