dabac_rs: Correct eval runner registration
The runner registration function always added a KVVec with capacity
`amount` for every missing entry, even though that is not necessary.
This means if a runner with a larger amount would register itself after
a runner with a smaller amunt, its allocation would be too small. Since
we access this with unsafe during performance eval, it is critical this
is correct.

Practically, all amounts will likely be the same during evaluation, but
this still needs to be correct.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>