abac: Allow loading larger object attributes
When the object attributes become too large, the write syscalls get
split up into multiple ones. This breaks the parser, which does not work
with partial data, and the interface does not wait for all data to be
written. This is because the buffers cannot grow as large as needed for
some data, mainly object representations for abac_trees and
abac_trees_anc, as they are allocated with kmalloc, which only hands out
full pages, which are limited to 4MB. To mitigate this, switch the
allocator to kvmalloc for all allocations in the fs, just to be sure.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>