dabac_rs: Fix handling of unset attributes
Accessing attributes that are not set should result in the rule being
evaluated to false and the access denied. Turns out that's not how the
code handles it, since it replaces the option way too early, which loses
the "unset attribute accessed" signal. Just returning false does not
work as further up there might be a disjunction where another
alternative evaluates to true, or even worse, there might be a negation
turning that very term true.

To fix this, carry the Option all the way to the top of the expression
evaluation, even if it makes the code a lot less nice. The conjunction
and disjunction evaluation now probably also does not short-curcuit
anymore unless a None is returned, but for the amount of time formula
resolution takes and the amount of time None values will be returned
during the evaluation, this is probably not the worst.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>