dabac_rs: abac: Do not lock perf results stores
In order to influence the perf evaluation as little as possible, do not
use locks to access the results store. This is possible, because each
runner has its own entry into the store. In C this is not UB (as far as
I know), so the lock can just not be taken.

In Rust this is a lot more tricky, multiple mutable references to the
same data structure (or static mut in this case) existing at the same
time is instant UB, so this uses only pointers until the Vec for each
runner is reached. At that point, access is exclusive, because each
runner can only send one request at a time. This is very delicate and
probably still UB in some way, I have not been able to test this under
miri. To me it feels similar to how `split_at_mut` works internally,
non-overlapping mutable references are ok, you just need to use pointers
to get there. So far the data I am getting looks good compared to what
the version with locks gave me, and the orchestrator also runs some
sanity checks on the data.

Another option would be one static mut for each runner, there are only
eight after all. This adds some branching though, where the result
depends on which user space process sent the request, so I am unsure if
the branch predictor would enjoy that. I might try that as well if I
still have time.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>