dabac_rs: Allow static_mut_refs in evaluation
The eval code accesses the perf results store without using locks during
evaluation, in order to slow down the code as little as possible. This
uses pointers until an exclusive mutable reference for a runner can be
created. The other uses of the store (initialization, results fetching)
are protected by a lock (which is hopefully safe from reordering).

Far from the safest interface, but so far it has provided results that
pass all sanity checks and show the same trends as the data gathered
using locks.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>