dabac_rs: Additionally store policy early returns
To also apply that for the evaluation, store the amount of times a
pre-condition resolution early returns because of unsatisfiable
attributions. These also become part of the result data fetched by the
orchestrator.

At some point there is a bit of a concern, how much collecting all this
additional data slows down the execution. The data I have on the time
difference between outer, inner and precise perf does not really
indicate that well.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>