dabac_rs: Avoid needless allocation to get filename
Previously, the file_get_full_name helper function would internally
allocate the buffer and return a CString. To do that, the result from
the FFI call even needs to be copied again, instead of just reusing the
orignal buffer, because the dentry_path_raw function confusingly does
not put the name at the beginning of the provided buffer for whatever
reason.

To prevent that, just have the caller provide the buffer, even if it
makes the interface less nice to use.

Signed-off-by: Lukas Fischer <kernel@o1oo11oo.de>